Simplified several things: 	-Now "Printable" objects can be streamed into XStr's using the regular << operators, in addition to calling their "print" method.  This means instead of 	str << "("; 	param->print(str); 	str << ")"; we can now write 	str << "("<<param<<")"; which is a good deal shorter and clearer.
	-XStr now has a copy constructor, so functions can
return XStrs.  This means instead of
	str << "void ";
	container->genBaseName(str);
	str << "("; ...
we can now write
	str << "void "<<container->baseName()<<"("; ...
which is a slight improvement.


	-Chare now has subclasses Array, Group, and Mainchare instead
of switching based on a magic "chareType" value.  This means we can
use virtual functions for the differences between these objects, instead
of ugly if(...) blocks.


These changes have resulted in about a 30% reduction in code size, and
(I think) dramatic improvement in readability in xi-symbol.C ..