Corrected handling of readonly messages. Readonly messages were earlier packed and never unpacked on PE0, sometimes they were deleted on PE 0.
Removed a race condition that caused net-smp version to hang after
initialization, the first PE on a node that reaches _initDone(), informs
it's peers on the same node that they should proceed to complete
initialization. This notification is in addition to the init Messages that arrive over the
network. The extra notification does not affect any other version.