Second major update of arrays: 	-Translator, support library changed to allow user-defined messages to array constructors (and multiple constructors). This significantly changed (and simplified!) the element creation sequence and translator-generated create code.
	-Translator now inherits *all* messages from CkArrayMessage.
This wastes a few bytes, but soon CkArrayMessage will be folded into
"envelope" anyway.

	-Array proxies now only accept a fixed type of index (1D, 2D, 3D,
user-defined), specified
in the .ci file.  This should prevent common indexing errors early.

	-Migration now uses much simpler, high-level "pup" library.
This will allow cross-platform pack/unpack, as well as checkpointing
to disk.


Major changes to:
	translator, array implementation

Minor changes to:
	tempo, header files

New: pack/unpack (pup) library.  I originally wrote this for Sameer's
checkpointing, but liked it so much I'm now using it for migration.
See pup.h or the manual for details.