Added "+no_outstanding_sends" option, which cleans out any pending sends each time we hit the scheduler.
Some (stupid) MPI implementations *require* MPI_Test
(or MPI_Wait) calls in order for MPI_Isend-sent messages
to actually make progress.

This can lead to the situation where messages are delayed
hundreds of milliseconds when a sender has other, non-MPI,
work to do just after a send.  A more reliable, but slower
fix would be to put the "while (MsgQueueLen>=0)" loop
right inside the CmiFreeSendFn; this version allows more
concurrency but may not be as fast as possible if, e.g.,
sends happen in the middle of an entry method.