Got rid of the ppn_factor hack. Now only one message will be sent to distinct processor at a time. So if two messages are sent to the same processor the second one will wait for the first one to be received. This will reduce the likely hood for stretched messages. I am also assuming the ONE_QXD_PER_PROCESSOR is set to 1 in the elan libraries. Using David Addison's library will enable this. Otherwise each message will wait for the previous messages envelope to be delivered.
Adding a queue for intermediate message sizes. So now there there are three queues, one for
small messages, one for intermediate sized messages and one for large messages which are
received in unexpected buffers.

Added functions to enable and disable blocking receives. When turned on PumpMsgs blocks if it
does not have any messages to receive. This gives the operating system time to schedule daemons.

Added a function to call hardware synchronous broadcast. This is a fast hardware broadcast which
uses pipelining and is aware of the underlying toplogy. FIXME: the elan broadcast is dumb though
asynchronous. It does not have pipelining. It uses binary trees as opposed to binomial trees. It
is also not node and topology aware.

CmiListSendFn: In Elan several messages can be sent from the same memory loation simulteneously.
So CmiListSendFn's do not copy the messages before sending and send them from the same location.
I also have sample (broken) code to copy the message into Elan memory which inproves the
bandwidth by 33% due to reduced DMA contention. It is commented out and I will commit it later.

 CVS: ----------------------------------------------------------------------