Added this very peculiar set of testing routines-- the ".tst" files are written in a bizarre dialect that can be preprocessed into either C++ (using the c_tst.h header) or f90 (using the f90_tst.h header).
The advantage to doing this is that we can write only
one set of testing code, and it automatically works for
C as well as F90, and is guaranteed to work the same way.
This should help prevent the common occurrence where one
or the other language binding is broken.