Added a simple implementation of the "memory aliasing" idea for thread stacks.  This is exactly like stack-copying threads, except the copy happens only in virtual address space-- the physical pages representing the stack never get moved.
This means memory aliased stacks:
   - Can be swapped very quickly (under 5us total penalty in AMPI)
   - Don't slow down with larger stacks
   - Don't consume large amounts of virtual address space like isomalloc stacks

Memory aliased stacks, however, will *NOT* work in SMP mode (because
all stacks get mapped to the same set of addresses), and
*will* break code that, e.g., swaps directly from one migratable
thread to the next without going through the scheduler, or creates
threads from inside a migratable thread (because the running stack
gets replaced with a new stack).

Because of the drawbacks, memory aliased stacks are off by default;
to enable them compile threads.c with -DCMK_THREADS_ALIAS_STACK=1.