These are MASSIVE changes that resulted from making the vmi-linux machine layer compatible with VMI 2.0 Beta 3 (and hopefully the final release).
To this end, I modified the way the VMI cache manager is used because there
are difficulties between this and the way Charm++ uses memory from within
its threads.  The advantage of this is that the code runs faster because
it pools memory through a single set of pools rather than having to call
the cache manager in a lot of different contexts.

Also, I have refactored the startup code in anticipation of making an
option to start up vmi-linux programs via a charmrun that understands the
concept of multiple clusters, so it can pay attention to what PEs appear
on one cluster and what PEs appear on a second cluster to try to minimize
the inter-cluster communications.