Added a "LDZeroLoads()" API routine, to clear out any stored load information.
This is useful if you don't want to turn instrumentation on and
off.  Manual instrumentation control can be difficult if you don't
know when you'll want to load balance-- it could be any time, but
it shouldn't include data from more than a few seconds ago.
Now we can just clear the loads every few seconds.