Fixed bug introduced by change of CMK_INCDIR and CMK_LIBDIR with OPTS_CXX and OPTS_LD respectively. They were originally omitted from the test scripts in configure.in, preventing the detection of desired system libraries like PAPI at build time. OPTS_CXX and OPTS_LD are now included alongside the original CMK_INCDIR and CMK_LIBDIR in configure.in, although I suspect the latter is not required because CMK_CXX and CMK_LD seem to usually include CMK_INCDIR and CMK_LIBDIR in their respective machine layers.
The configure.in script has also been fixed to only test for PAPI_TOT_CYC
which applies to all machine architectures. The other test for PAPI_FP_INS
does not apply to the itanium 2 architecture.

Also updated the tracing module such that papiEventSet is initialized to
PAPI_NULL as mandated by the PAPI 3 specification.