Modified vmi-linux machine layer to understand the concept of distinct clusters.  Now it is possible for a PE to determine whether it is in the same cluster as any other PE.  This will be necessary for the work coming up in March.
The user may specify which cluster a given sub-job belongs to, or may tell
the machine layer to probe the PE-to-cluster mapping automatically.  In
order to do this, the PEs each measure the point-to-point latency to all
other PEs in the computation and then report this data to PE 0.  PE 0
collects all the latency data and uses it to compute the cluster map.
It then distributes this map to all PEs in the computation.

The latency collection routines are also important for Grid load balancer
work that will be checked in later.