My implementation of eager RDMA message sends has to be able to intercept calls to free() on the discarded message buffer and instead of freeing the memory send a "update credit" message to the sender to mark the buffer as available for more messages.  To do this, I prepend the standard Charm memory header (size field, ref field) with an additional field which is a pointer to the eager handle.  For a CmiFree() this field is checked and if it points to a handle, the handle is updated; otherwise if NULL the memory is just a regular allocation that can be freed.
The problem with this is when using some of Sameer's comlib routines
like StreamingStrategy and MeshStreamingStrategy (which I originally wrote
but which Sameer modified to include this bug here).  These strategies
pack messages together and send a single stream containing multiple
packed messages.  In order to do this, he uses the ref fields for the
messages in a packed buffer, making them point back to the ref field of
the larger packed buffer.  Since he transfers the entire buffer of packed
memory and does not know about my additional prepended field, he tromples
my marker.

I think the Converse commbench routines also exercise this fault.

To correct this, I modified my code to disable the eager send protocol
by default, so this additional memory field is never used.  If the
environment variable CMI_VMI_EAGER_PROTOCOL=1 is set, eager sends can be
turned on as usual and the additional memory field is added to all memory
allocations.

My machine layer now passes autobuild.