Updated many objects to use initializers. Tweaked UnsetTS logic to match use of POSE_UnsetTS instead of a mix of that and -1.
Changed value of POSE_UnsetTS from -1 to -20, so misuse of uninitialized
timestamps is more likely to trigger an assert failue.

This version should be slightly faster than the old version.  It is notably
more correct in that BigNetSim is now processing large trace runs in
parallel when using LONG_TIMESTAMPS.