AMPI: Add GPU support via AMPI_GPU_invoke/MPI_Wait
Integrate support for Lukasz's GPU manager API through normal-looking
AMPI functions. The client code sets up a work request the same as it
would in Charm++, but doesn't set a callback. Passing this work
request to AMPI_GPU_invoke enqueues it for the GPU and gives back an
MPI_Request. The client code can MPI_Wait on this MPI_Request when it
needs to ensure completion.