Reductions: Set the reference number on reduction result messages
SDAG's `when' construct supports a notion of a tag, or `reference
number' on incoming messages to restrict processing order. On normal
messages, this is set with CkSetRefNum, and parameter-marshalled
messages use the first integer parameter.

Pritish found that it would be useful if the messages generated by
reductions had a predictable reference number. Since the user code
generally doesn't handle reduction messages until it's too late to be
useful, set the reference number of the message handed to the
reduction callback to the reduction number.