configure: Drop archaic test on C library version
Years ago, there were versions of the GNU C library that didn't support
the pthreads API. Since SMP builds needed that, we tested for it.
Those versions are long gone, and the test is failing on modern (>2.10)
versions of glibc. Rather than try to fix it, just dump the test.