Reductions: Set the reference number on result messages to the userFlag
Setting the reference number on the CkReductionMsg passed to the
callback to the reduction number was a dirty hack to allow Pritish to
get something predictable in the SDAG code Charisma generated. See
74e51ab49dc766f93bd3c297254ffb157294ae3a about that.

To expose this as reliable, controllable functionality, pick up the
caller-provided userFlag value (defaults to -1) and use that instead.