CProxySection_ArrayBase: Fix bug caused by inconsistent new/deletes
Implementing cross-array-sections required new constructors for creating multi-array
sections. These constructors always used new[] even if the section spanned just one
chare array. The destructor on the other hand uses delete / delete[] depending on the
number of arrays in the section. Using these constructors to create a single-array
section causes segfaults in the destructor.

Quick fix so that new[] is used only if _nsid > 1 irrespective of which constructor
is invoked.