Added 'unused' attribute to idx2str to silence g++ warnings.
'g++ -Wall' warns any time you include charm++.h because idx2str is a
static function that is defined but not used. I added a macro to mark
the function with an attribute that silences the warning (for g++ only;
it has no effect unless __GNUC__ is defined). This allows charm programs
to be compiled with -Wall without generating any warnings.