Add a '-production' flag for architecture-specific features
Abhinav requested a flag to indicate that architecture-specific
production-mode features should be enabled. Add it from front (/build)
to back (charmc), with documentation.

To add something to the compiler flags, add an appropriate definition
of CMK_PRODUCTION or CMK_{C,CXX,F90,LD,LDXX}_PRODUCTION to
conv-mach.sh or cc-foo.sh, and the -production flag will include it in
the options passed to the underlying compiler.

I did *not* make a test for this in the configure script, nor define a
C preprocessor macro CMK_PRODUCTION, because code should never be
testing that directly. Architecture-specific scripts should be
enabling particular features where they're appropriate.