CentralLB: remove redundant parameter 'count'
Remove the 'count' parameter from four functions in CentralLB and fix the
corresponding compilation errors in combination LB implementations. The next
step is to change the work implementations in other load balancers.

Change docs and version number to reflect the ldb API change, so that
application load balancers in NAMD and ChaNGa can work either way.